:bug: fix file proccess bug

+++ src/utils/file.ts
@@ -13,13 +13,17 @@ export const checkDirAndCreate = (filePath: string) => {
 }
 
 export const removeFile = (filePath: string) => {
-  fs.unlink(filePath, (err) => {
-    if (err) throw err
-  })
+  if (fs.existsSync(filePath)) {
+    fs.unlink(filePath, (err) => {
+      if (err) throw err
+    })
+  }
 }
 
 export const removeFileDir = (filePath: string) => {
-  fs.rmdir(filePath, (err) => {
-    if (err) throw err
-  })
+  if (fs.existsSync(filePath)) {
+    fs.rmdir(filePath, (err) => {
+      if (err) throw err
+    })
+  }
 }
